LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for time input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, min, max, [keySequence]).
Normal cases
PASS stepUp("07:13", null, null, null) is "07:14"
PASS stepDown("07:13", null, null, null) is "07:12"
PASS test("07:13", null, null, null, ["upArrow", "upArrow", "upArrow"]) is "07:16"
PASS test("07:13", null, null, null, ["downArrow", "downArrow", "downArrow"]) is "07:10"
Milliseconds
PASS stepUp("07:13:00.000", 0.001, null, null) is "07:13:00.001"
PASS stepDown("07:13:00.001", 0.001, null, null) is "07:13"
PASS stepUp("07:13:00.999", 0.001, null, null) is "07:13"
PASS stepDown("07:13:00.000", 0.001, null, null) is "07:13:00.999"
PASS stepUp("07:13:00.000", 0.250, null, null) is "07:13:00.250"
PASS stepDown("07:13:00.500", 0.250, null, null) is "07:13:00.250"
PASS stepUp("07:13:00.750", 0.250, null, null) is "07:13"
PASS stepDown("07:13:00.000", 0.250, null, null) is "07:13:00.750"
PASS stepUp("07:13:00.099", 0.100, null, null) is "07:13:00.100"
PASS stepDown("07:13:00.101", 0.100, null, null) is "07:13:00.100"
PASS stepUp("07:13:00.500", 0.0001, null, null) is "07:13:00.501"
PASS stepDown("07:13:00.500", 0.0001, null, null) is "07:13:00.499"
PASS stepUp("07:13:00.500", 0, null, null) is "07:13"
PASS stepDown("07:13:00.500", 0, null, null) is "07:13"
PASS stepUp("07:13:00.500", 60, null, null) is "07:13"
PASS stepDown("07:13:00.500", 60, null, null) is "07:13"
PASS stepUp("07:13:00.500", 0.100, "00:00:00.050", null) is "07:13:00.550"
PASS stepDown("07:13:00.500", 0.100, "00:00:00.050", null) is "07:13:00.450"
PASS test("07:13:00.500", 0.100, null, null, ["delete", "upArrow"]) is "07:13"
PASS test("07:13:00.500", 0.100, null, null, ["delete", "downArrow"]) is "07:13:00.900"
PASS test("07:13:00.500", 0.100, "00:00:00.050", null, ["delete", "upArrow"]) is "07:13:00.050"
PASS test("07:13:00.500", 0.100, "00:00:00.050", null, ["delete", "downArrow"]) is "07:13:00.950"
Seconds
PASS stepUp("07:13:00", 1, null, null) is "07:13:01"
PASS stepDown("07:13:01", 1, null, null) is "07:13"
PASS stepUp("07:13:59", 0.001, null, null) is "07:13"
PASS stepDown("07:13:00", 0.001, null, null) is "07:13:59"
PASS stepUp("07:13:00", 10, null, null) is "07:13:10"
PASS stepDown("07:13:30", 10, null, null) is "07:13:20"
PASS stepUp("07:13:50", 10, null, null) is "07:13"
PASS stepDown("07:13:00", 10, null, null) is "07:13:50"
PASS stepUp("07:13:09", 10, null, null) is "07:13:10"
PASS stepDown("07:13:11", 10, null, null) is "07:13:10"
PASS stepUp("07:13:30", 2.5, null, null) is "07:13:31"
PASS stepDown("07:13:30", 2.5, null, null) is "07:13:29"
PASS stepUp("07:13:30", 0, null, null) is "07:13"
PASS stepDown("07:13:30", 0, null, null) is "07:13"
PASS stepUp("07:13:30", 3600, null, null) is "07:13"
PASS stepDown("07:13:30", 3600, null, null) is "07:13"
PASS stepUp("07:13:30", 11, null, null) is "07:13:31"
PASS stepDown("07:13:30", 11, null, null) is "07:13:29"
PASS stepUp("07:13:30", 10, "00:00:05", null) is "07:13:35"
PASS stepDown("07:13:30", 10, "00:00:05", null) is "07:13:25"
PASS test("07:13:30", 10, null, null, ["delete", "upArrow"]) is "07:13"
PASS test("07:13:30", 10, null, null, ["delete", "downArrow"]) is "07:13:50"
PASS test("07:13:30", 10, "00:00:05", null, ["delete", "upArrow"]) is "07:13:05"
PASS test("07:13:30", 10, "00:00:05", null, ["delete", "downArrow"]) is "07:13:55"
Minutes
PASS stepUp("07:13", 1, null, null) is "07:14"
PASS stepDown("07:13", 1, null, null) is "07:12"
PASS stepUp("07:59", 0.001, null, null) is "07:00"
PASS stepDown("07:00", 0.001, null, null) is "07:59"
PASS stepUp("07:00", 600, null, null) is "07:10"
PASS stepDown("07:10", 600, null, null) is "07:00"
PASS stepUp("07:50", 600, null, null) is "07:00"
PASS stepDown("07:00", 600, null, null) is "07:50"
PASS stepUp("07:13", 600, null, null) is "07:20"
PASS stepDown("07:13", 600, null, null) is "07:10"
PASS stepUp("07:13", 150, null, null) is "07:14"
PASS stepDown("07:13", 150, null, null) is "07:12"
PASS stepUp("07:13", 0, null, null) is "07:14"
PASS stepDown("07:13", 0, null, null) is "07:12"
PASS stepUp("07:13", 36000, null, null) is "07:00"
PASS stepDown("07:13", 36000, null, null) is "07:00"
PASS stepUp("07:13", 660, null, null) is "07:14"
PASS stepDown("07:13", 660, null, null) is "07:12"
PASS stepUp("07:13", 600, "00:05", null) is "07:15"
PASS stepDown("07:13", 600, "00:05", null) is "07:05"
PASS test("07:13", 600, null, null, ["delete", "upArrow"]) is "07:00"
PASS test("07:13", 600, null, null, ["delete", "downArrow"]) is "07:50"
PASS test("07:13", 600, "00:05", null, ["delete", "upArrow"]) is "07:05"
PASS test("07:13", 600, "00:05", null, ["delete", "downArrow"]) is "07:55"
Hours, 1-12
PASS stepUp("07:00", 1, null, null) is "08:00"
PASS stepDown("07:00", 1, null, null) is "06:00"
PASS stepUp("11:00", 1, null, null) is "00:00"
PASS stepDown("00:00", 1, null, null) is "11:00"
PASS stepUp("06:00", 7200, null, null) is "08:00"
PASS stepDown("06:00", 7200, null, null) is "04:00"
PASS stepUp("10:00", 7200, null, null) is "00:00"
PASS stepDown("00:00", 7200, null, null) is "10:00"
PASS stepUp("07:00", 7200, null, null) is "08:00"
PASS stepDown("07:00", 7200, null, null) is "06:00"
PASS stepUp("06:00", 3601, null, null) is "07:00"
PASS stepDown("06:00", 3601, null, null) is "05:00"
PASS stepUp("06:00", 0, null, null) is "07:00"
PASS stepDown("06:00", 0, null, null) is "05:00"
PASS stepUp("06:00", 86400, null, null) is "00:00"
PASS stepDown("06:00", 86400, null, null) is "00:00"
PASS stepUp("06:00", 36000, null, null) is "07:00"
PASS stepDown("06:00", 36000, null, null) is "05:00"
PASS stepUp("06:00", 7200, "01:00", null) is "07:00"
PASS stepDown("06:00", 7200, "01:00", null) is "05:00"
PASS test("06:00", 7200, null, null, ["delete", "upArrow"]) is "02:00"
PASS test("06:00", 7200, null, null, ["delete", "downArrow"]) is "00:00"
PASS getUserAgentShadowTextContent(input) is "12:00 AM"
PASS test("06:00", 7200, "01:00", null, ["delete", "upArrow"]) is "01:00"
PASS test("06:00", 7200, "01:00", null, ["delete", "downArrow"]) is "11:00"
Hours, 0-23
The tests in this block fail on platforms without the lang-attribute-aware-form-control-UI feature.
PASS stepUp("07:00", 1, null, null) is "08:00"
PASS stepDown("07:00", 1, null, null) is "06:00"
PASS stepUp("23:00", 1, null, null) is "00:00"
PASS stepDown("00:00", 1, null, null) is "23:00"
PASS stepUp("06:00", 7200, null, null) is "08:00"
PASS stepDown("06:00", 7200, null, null) is "04:00"
PASS stepUp("22:00", 7200, null, null) is "00:00"
PASS stepDown("00:00", 7200, null, null) is "22:00"
PASS stepUp("07:00", 7200, null, null) is "08:00"
PASS stepDown("07:00", 7200, null, null) is "06:00"
PASS stepUp("06:00", 3601, null, null) is "07:00"
PASS stepDown("06:00", 3601, null, null) is "05:00"
PASS stepUp("06:00", 0, null, null) is "07:00"
PASS stepDown("06:00", 0, null, null) is "05:00"
PASS stepUp("06:00", 86400, null, null) is "00:00"
PASS stepDown("06:00", 86400, null, null) is "00:00"
PASS stepUp("06:00", 36000, null, null) is "07:00"
PASS stepDown("06:00", 36000, null, null) is "05:00"
PASS stepUp("06:00", 7200, "01:00", null) is "07:00"
PASS stepDown("06:00", 7200, "01:00", null) is "05:00"
PASS test("06:00", 7200, null, null, ["delete", "upArrow"]) is "00:00"
PASS getUserAgentShadowTextContent(input) is "00:00"
PASS test("06:00", 7200, null, null, ["delete", "downArrow"]) is "22:00"
PASS test("06:00", 7200, "01:00", null, ["delete", "upArrow"]) is "01:00"
PASS test("06:00", 7200, "01:00", null, ["delete", "downArrow"]) is "23:00"

PASS successfullyParsed is true

TEST COMPLETE
